Added Basic Errors and Parser
Refactored code into individual files

Added re exports

Added custom errors

Added outline (helper functions) of a parser.
